Outline how to share complex code between old and new style controllers.
More use of Pydantic models in the old code and more use of "view" functions below the controller code - mirroring inn some ways what is done for dependency management APIs (which I had structured to be very lightweight in an alternative framework such as this).
