+++ lib/galaxy/managers/jobs.py
@@ -1,7 +1,12 @@
 import json
 import logging
+import typing
 
 from boltons.iterutils import remap
+from pydantic import (
+    BaseModel,
+    Field,
+)
 from sqlalchemy import and_, false, func, or_
 from sqlalchemy.orm import aliased
 from sqlalchemy.sql import select
@@ -23,6 +28,10 @@ from galaxy.util import (
 )
 
 log = logging.getLogger(__name__)
+
+
+class JobLock(BaseModel):
+    active: bool = Field(title="Job lock status", description="If active, jobs will not dispatch")
 
 
 def get_path_key(path_tuple):
@@ -49,6 +58,13 @@ class JobManager:
     def __init__(self, app):
         self.app = app
         self.dataset_manager = DatasetManager(app)
+
+    def job_lock(self):
+        return JobLock(active=self.app.job_manager.job_lock)
+
+    def update_job_lock(self, job_lock: JobLock):
+        self.app.queue_worker.send_control_task('admin_job_lock', kwargs={'job_lock': job_lock.active}, get_response=True)
+        return self.job_lock()
 
     def get_accessible_job(self, trans, decoded_job_id):
         job = trans.sa_session.query(trans.app.model.Job).filter(trans.app.model.Job.id == decoded_job_id).first()
@@ -366,6 +382,30 @@ class JobSearch:
             return job
         log.info("No equivalent jobs found %s", search_timer)
         return None
+
+
+def view_show_job(trans, job, full: bool) -> typing.Dict:
+    is_admin = trans.user_is_admin
+    job_dict = trans.app.security.encode_all_ids(job.to_dict('element', system_details=is_admin), True)
+    if full:
+        job_dict.update(dict(
+            tool_stdout=job.tool_stdout,
+            tool_stderr=job.tool_stderr,
+            job_stdout=job.job_stdout,
+            job_stderr=job.job_stderr,
+            stderr=job.stderr,
+            stdout=job.stdout,
+            job_messages=job.job_messages
+        ))
+
+        if is_admin:
+            if job.user:
+                job_dict['user_email'] = job.user.email
+            else:
+                job_dict['user_email'] = None
+
+            job_dict['job_metrics'] = summarize_job_metrics(trans, job)
+    return job_dict
 
 
 def invocation_job_source_iter(sa_session, invocation_id):
